feat: separando configurações de estilo
